services/forecaster/requirements.txt
Flask==3.0.0
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
scikit-learn==1.5.1
google-cloud-bigquery==3.25.0
pyarrow==16.1.0
gunicorn==21.2.0
prophet==1.1.5
cmdstanpy==1.2.4
gcloud projects list
